ai-service/main.py
from flask import Flask, jsonify, request
import os
from dotenv import load_dotenv

load_dotenv()

app = Flask(__name__)

@app.route('/health', methods=['GET'])
def health():
    return jsonify({'status': 'OK', 'service': 'ai-service'})

@app.route('/api/v1/analyze', methods=['POST'])
def analyze():
    """Analyze teacher performance and quality"""
    data = request.json
    
    # TODO: Implement sentiment analysis
    # TODO: Implement performance scoring
    # TODO: Implement risk detection
    # TODO: Implement recommendations engine
    
    return jsonify({
        'status': 'pending',
        'analysis_id': 1,
        'message': 'Analysis queued for processing'
    })

@app.route('/api/v1/predict', methods=['POST'])
def predict():
    """Predict teacher performance metrics"""
    data = request.json
    
    # TODO: Implement ML models for prediction
    
    return jsonify({
        'prediction': None,
        'confidence': 0.0
    })

if __name__ == '__main__':
    port = os.getenv('FLASK_PORT', 5000)
    app.run(debug=True, port=int(port))
