ai-service/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
openai==0.27.8
pandas==2.0.3
